% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatedata.R
\name{generatedata}
\alias{generatedata}
\title{generatedata}
\usage{
generatedata(nc, ng, xm, xsigma, ym, ysigma, p1, p2)
}
\arguments{
\item{nc}{number of cells}

\item{ng}{number of genes}

\item{xm}{low-dimensional mean of x-values}

\item{xsigma}{low-dimensional sd of x-values}

\item{ym}{low-dimensional mean of y-values}

\item{ysigma}{low-dimensional sd of y-values}

\item{p1}{batch 1 cell-type proportions}

\item{p2}{batch 2 cell-type proportions}
}
\description{
generatedata
}
