% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTSNE.R
\name{makeTSNE}
\alias{makeTSNE}
\title{Generate a tSNE Plot to visually evaluate the quality of batch correction methods}
\usage{
makeTSNE(seed = 0, ncells = 1000, ngenes = 100, xmus = c(0, 5, 5),
  xsds = c(1, 0.1, 1), ymus = c(5, 5, 0), ysds = c(1, 0.1, 1),
  prop1 = c(0.3, 0.5, 0.2), prop2 = c(0.65, 0.3, 0.05))
}
\arguments{
\item{seed}{seed for reproducibility}

\item{ncells}{The number of cells in the sequencing experiment.}

\item{ngenes}{The number of genes per cell.}

\item{xmus}{A vector of three numbers, representing the x-coordinates of the three cell types in a two-dimensional space before projection to the space of all genes.}

\item{xsds}{The standard deviations of x-coordinates of the three cell types in a two-dimensional space before projection to the space of all genes.}

\item{ymus}{A vector of three numbers, representing the y-coordinates of the three cell types in a two-dimensional space before projection to the space of all genes.}

\item{ysds}{The standard deviations of y-coordinates of the three cell types in a two-dimensional space before projection to the space of all genes.}

\item{prop1}{A vector of three numbers, representing the proportions of the three cell types in the first batch. The elements of the vector must sum to 1.}

\item{prop2}{A vector of three numbers, representing the proportions of the three cell types in the second batch. The elements of the vector must sum to 1.}
}
\value{
Returns four tSNE plots (on the same figure).
}
\description{
Given a set of parameters, this function will simulate data and then output four tSNE plots: one corresponding to clustering after no batch correction, and another three, each corresponding to clustering after one of the three batch correction methods "limma," "combat," and "mnn."
}
\examples{
\dontrun{
makeTSNE(
seed=0,ncells=1000,ngenes=100,
xmus=c(0,5,5),xsds=c(1,0.1,1),
ymus=c(5,5,0),ysds=c(1,0.1,1),
prop1=c(0.3,0.5,0.2),prop2=c(0.65,0.3,0.05)
)
}
}
