% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mystudy.R
\docType{data}
\name{mystudy}
\alias{mystudy}
\title{mystudy}
\usage{
data(mystudy)
}
\description{
The object of simulation results returned from the "simstudy" function when run on the parameter combinations from the tutorial for the mnnsim package.
}
