% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docluster.R
\name{docluster}
\alias{docluster}
\title{docluster}
\usage{
docluster(mydata, type = "uncorrected", silscores = TRUE,
  cosnorm = TRUE)
}
\arguments{
\item{mydata}{output from generatedata function}

\item{type}{batch correction method to use before clustering}

\item{silscores}{whether to produce silhoutte scores or not}

\item{cosnorm}{whether to perform cosine normalization}
}
\description{
docluster
}
