% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mystudy.R
\docType{data}
\name{mystudy}
\alias{mystudy}
\title{mystudy}
\description{
The data object consisting of simulation results returned from the "simstudy" function in the package tutorial for the mnnsim package.
}
\examples{
 \dontrun{data(mystudy)
simresults(finishedstudy=mystudy,sds=T)
}
}
